refactor: PraPrice → MarketPrice로 이름 변경하여 Detailed로
기존의 시세페이지인 MarketPrice는 PraPrice를 div하나로 더 감싸줄 뿐,
결국 PraPrice가 시세페이지와 다름이 없다.

때문에 기존의 MarketPrice.tsx를 삭제하고,
PraPrice를 MarketPrice로 변경하여 Detailed 디렉토리로 이동
